fix: moved wipeout only when necessary or forced by database checker

+++ engine/src/main/java/com/arcadedb/engine/LocalBucket.java
@@ -910,8 +910,7 @@ public class LocalBucket extends PaginatedComponent implements Bucket {
 
     MutablePage page = null;
     try {
-      page = database.getTransaction()
-          .getPageToModify(new PageId(database, file.getFileId(), pageId), pageSize, false);
+      page = database.getTransaction().getPageToModify(new PageId(database, file.getFileId(), pageId), pageSize, false);
 
       final short recordCountInPage = page.readShort(PAGE_RECORD_COUNT_IN_PAGE_OFFSET);
       if (positionInPage >= recordCountInPage)
@@ -1022,7 +1021,7 @@ public class LocalBucket extends PaginatedComponent implements Bucket {
     }
   }
 
-  public void compressPage(final MutablePage page) throws IOException {
+  public void compressPage(final MutablePage page, final boolean forceWipeOut) throws IOException {
     final short recordCountInPage = page.readShort(PAGE_RECORD_COUNT_IN_PAGE_OFFSET);
 
     final List<int[]> orderedRecordContentInPage = getOrderedRecordsInPage(page, recordCountInPage, false);
@@ -1064,7 +1063,9 @@ public class LocalBucket extends PaginatedComponent implements Bucket {
       }
 
     }
-    wipeOutFreeSpace(page, recordCountInPage);
+
+    if (forceWipeOut || !holes.isEmpty())
+      wipeOutFreeSpace(page, recordCountInPage);
   }
 
   private void wipeOutFreeSpace(final MutablePage page, final short recordCountInPage) throws IOException {
@@ -1076,8 +1077,7 @@ public class LocalBucket extends PaginatedComponent implements Bucket {
         getFreeSpaceInPage(pageAnalysis);
 
         if (pageAnalysis.spaceAvailableInCurrentPage > 0)
-          page.writeZeros(pageAnalysis.newRecordPositionInPage,
-              page.getMaxContentSize() - pageAnalysis.newRecordPositionInPage);
+          page.writeZeros(pageAnalysis.newRecordPositionInPage, page.getMaxContentSize() - pageAnalysis.newRecordPositionInPage);
       } catch (Exception e) {
         // IGNORE IT
         LogManager.instance().log(this, Level.SEVERE, "Error on wiping out page content", e);
@@ -1194,9 +1194,8 @@ public class LocalBucket extends PaginatedComponent implements Bucket {
           continue;
         }
       } catch (Exception e) {
-        LogManager.instance()
-            .log(this, Level.SEVERE, "Error on loading record #" + fileId + ":" + (page.pageId.getPageNumber() * maxRecordsInPage)
-                + positionInPage);
+        LogManager.instance().log(this, Level.SEVERE,
+            "Error on loading record #" + fileId + ":" + (page.pageId.getPageNumber() * maxRecordsInPage) + positionInPage);
         continue;
       }
 
@@ -1243,8 +1242,8 @@ public class LocalBucket extends PaginatedComponent implements Bucket {
         // AVOID INFINITE LOOP?
         LogManager.instance().log(this, Level.SEVERE,
             "Infinite loop on loading multi-page record " + originalRID + " chunk " + chunkPageId + "/" + chunkPositionInPage);
-        throw new DatabaseOperationException("Infinite loop on loading multi-page record " + originalRID + " chunk "
-            + chunkPageId + "/" + chunkPositionInPage);
+        throw new DatabaseOperationException(
+            "Infinite loop on loading multi-page record " + originalRID + " chunk " + chunkPageId + "/" + chunkPositionInPage);
       }
 
       page = nextPage;
@@ -1303,8 +1302,7 @@ public class LocalBucket extends PaginatedComponent implements Bucket {
 
       final int spaceNeededForChunk = bufferSize + 2 + INT_SERIALIZED_SIZE + LONG_SERIALIZED_SIZE;
 
-      final PageAnalysis pageAnalysis = findAvailableSpace(currentPage.pageId.getPageNumber(), spaceNeededForChunk,
-          txPageCounter);
+      final PageAnalysis pageAnalysis = findAvailableSpace(currentPage.pageId.getPageNumber(), spaceNeededForChunk, txPageCounter);
       if (!pageAnalysis.createNewPage) {
         nextPage = database.getTransaction().getPageToModify(pageAnalysis.page.pageId, pageSize, false);
         newPosition = pageAnalysis.newRecordPositionInPage;
@@ -1705,9 +1703,8 @@ public class LocalBucket extends PaginatedComponent implements Bucket {
    * 3. if the tree map is full (size > MAX_PAGES_GATHER_STATS), stop
    */
   public void gatherPageStatistics() {
-    if (timeOfLastStats == 0L ||
-        (System.currentTimeMillis() - timeOfLastStats > MAX_TIMEOUT_GATHER_STATS && changesFromLastStats > 0)
-    )
+    if (timeOfLastStats == 0L || (System.currentTimeMillis() - timeOfLastStats > MAX_TIMEOUT_GATHER_STATS
+        && changesFromLastStats > 0))
       try {
         int txPageCount = getTotalPages();
